requirements.txt
altgraph==0.17.3
asgiref==3.5.2
autopep8==2.0.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-baton==2.6.0
fontawesomefree==6.2.1
google-api-core==2.11.0
google-api-python-client==2.68.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.57.0
httplib2==0.21.0
idna==3.4
oauth2client==1.5.2
protobuf==4.21.10
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
pyparsing==3.0.9
requests==2.28.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.13
